requirements.txt
fastapi==0.61.2
psycopg2-binary==2.8.6
requests==2.25.1
uvicorn==0.12.3
